pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = [
    {name = "yulichka.metel",email = "yulichka.metel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mypy (>=1.15.0,<2.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "pytest-assume (>=2.4.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = 3.13
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = true
